Update README.md
I present a comprehensive study on proposed TTAG-Net “topological attention graphical model for road semantic segmentation with Gating signals on network middle-block and attention-based mechanism on decoder side for spatial resolution recovery, mainly focusing on mainly feature to extract. We this model to serve and stand out of the crowd in deep learning and geospatial or spatial planning field. The findings have the potential to advance our understanding of how multimodal fusion and can be applied in various practical applications such as large-scale land cover mapping